+++ src/renderer/actions/apps/mobile/adb-commands.ts
@@ -77,15 +77,6 @@ export async function hasCertInstalled(
   }
 }
 
-export function stringAsStream(input: string) {
-  const contentStream = new stream.Readable();
-  // eslint-disable-next-line no-underscore-dangle
-  contentStream._read = () => {};
-  contentStream.push(input);
-  contentStream.push(null);
-  return contentStream;
-}
-
 export const checkProxy = async (adbClient: adb.Client, deviceId: string) => {
   return adbClient
     .getDevice(deviceId)
