+++ src/renderer/actions/apps/mobile/utils.ts
@@ -14,20 +14,20 @@ export function delay(durationMs: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, durationMs));
 }
 
-export async function waitUntil<T extends unknown>(
-  delayMs: number,
-  tries: number,
-  test: () => Promise<T>
-): Promise<Exclude<T, false>> {
-  let result = tries > 0 && (await test());
+// export async function waitUntil<T extends unknown>(
+//   delayMs: number,
+//   tries: number,
+//   test: () => Promise<T>
+// ): Promise<Exclude<T, false>> {
+//   let result = tries > 0 && (await test());
 
-  while (tries > 0 && !result) {
-    // eslint-disable-next-line no-param-reassign
-    tries -= 1;
-    await delay(delayMs);
-    result = await test();
-  }
+//   while (tries > 0 && !result) {
+//     // eslint-disable-next-line no-param-reassign
+//     tries -= 1;
+//     await delay(delayMs);
+//     result = await test();
+//   }
 
-  if (!result) throw new Error(`Wait loop failed`);
-  else return result as Exclude<T, false>;
-}
+//   if (!result) throw new Error(`Wait loop failed`);
+//   else return result as Exclude<T, false>;
+// }
